requirements.txt
cachetools==5.3.0 ; python_full_version >= "3.8.1" and python_version < "4.0"
certifi==2022.12.7 ; python_full_version >= "3.8.1" and python_version < "4"
cfgv==3.3.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
charset-normalizer==3.1.0 ; python_full_version >= "3.8.1" and python_version < "4"
distlib==0.3.6 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
filelock==3.12.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
flake8==6.0.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
google-api-core==2.11.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
google-api-core[grpc]==2.11.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
google-api-python-client==2.80.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
google-auth-httplib2==0.1.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
google-auth==2.16.2 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
google-cloud-bigquery==3.6.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
google-cloud-core==2.3.2 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
google-cloud-pubsub==2.21.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
google-crc32c==1.5.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
google-resumable-media==2.4.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
googleapis-common-protos==1.58.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
googleapis-common-protos[grpc]==1.58.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
grpc-google-iam-v1==0.13.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
grpcio-status==1.51.3 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
grpcio==1.51.3 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
httplib2==0.21.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
identify==2.5.24 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
idna==3.4 ; python_full_version >= "3.8.1" and python_version < "4"
isort==5.12.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
mccabe==0.7.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
nodeenv==1.8.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
packaging==23.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
platformdirs==3.5.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pre-commit==3.3.2 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
proto-plus==1.22.2 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
protobuf==4.22.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pyasn1-modules==0.2.8 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pyasn1==0.4.8 ; python_full_version >= "3.8.1" and python_version < "4"
pycodestyle==2.10.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pyflakes==3.0.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pyparsing==3.0.9 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
python-dateutil==2.8.2 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
pyyaml==6.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
requests==2.28.2 ; python_full_version >= "3.8.1" and python_version < "4"
rsa==4.9 ; python_full_version >= "3.8.1" and python_version < "4"
setuptools==67.8.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
six==1.16.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
uritemplate==4.1.1 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
urllib3==1.26.14 ; python_full_version >= "3.8.1" and python_version < "4"
virtualenv==20.23.0 ; python_full_version >= "3.8.1" and python_full_version < "4.0.0"
